merging feature 4

+++ src/main/java/edu/fiu/group24/group_24_project/Books/BookRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface BookRepository extends JpaRepository<Book, String> {
     //I KNOW THIS NAME IS WEIRD BUT ITS LIKE THIS BECAUSE OF HOW NAMING CONVENTIONS WORK
     //THE METHOD NAME LIKE THIS TELLS THE CODE TO LOOK FOR "bookAuthor" (name of the local variable) VIA THE "author_id"
-    List<Book> findByBookAuthor_AuthorId(Integer author_id);
+    List<Book> findByBookAuthor_AuthorId(long author_id);
 }
